apps/streamlit/pyproject.toml
[project]
name = "googly-streamlit"
version = "0.1.0"
description = "Googly — Streamlit prototype UI"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.42.0",
    "agno~=2.5.0",
    "openai>=2.16.0",
    "ensembledata>=0.2.11",
    "httpx~=0.28.0",
    "psycopg[binary]>=3.1.0",
    "python-dotenv>=1.0.0",
    "PyJWT~=2.10.0",
    "structlog>=24.0.0",
    "sqlalchemy~=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
